requirements.txt
Flask==1.0.2
gunicorn==19.9.0
beautifulsoup4==4.6.3
line-bot-sdk==1.8.0
imgurpython==1.1.7
pymongo==3.9.0
pandas==0.25.2
lxml==4.4.1
APScheduler==3.0.0
flask_apscheduler==1.11.0